% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdsdata-package.R
\docType{package}
\name{bdsdata-package}
\alias{bdsdata}
\alias{bdsdata-package}
\title{bdsdata: Data sets for the course: Behavioral Data Science I}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Nicholas Duran \email{nicholas.duran@asu.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
